api/car/model/serializers.py
from rest_framework import serializers

from apps.car.models import CarModel


class CarModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = CarModel
        fields = '__all__'
