requirements.txt
Django==3.1.7
djangorestframework==3.12.4
Pillow==8.1.2
pyyaml==5.4.1
uritemplate==3.0.1
